+++ examples/resources/iosxe_nat/resource.tf
@@ -1,0 +1,13 @@
+resource "iosxe_nat" "example" {
+  inside_source_interface_lists = [
+    {
+      id = "10"
+      interfaces = [
+        {
+          interface = "GigabitEthernet10"
+          overload  = true
+        }
+      ]
+    }
+  ]
+}